+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @formwright/ai
 
+## 0.2.2
+
+### Patch Changes
+
+- @formwright/schema@0.2.2
+
+## 0.2.1
+
+### Patch Changes
+
+- @formwright/schema@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes